fix: moved function to different folder and changed logs
